feat(#339): refactor shared project mutation hooks and move leave project logic to separate file
